Secure crates can be made tamper proof now
This will deter players from abusing projectiles to open boxes, as there
is now a chance to destroy the contents and cause a (small) fire, or a
small explosion that will alert security.

Crates will not detonate if unlocked so the CE doesn't have their life
made hell for using an emitter to unlock the supermatter core.