Fixes #568 again
Also tweaked spawning gear for CentCom observers, and fixed some
problems in the office. Also removed the CentCom stamp so it must be
spawned with instead.

The idea of a CentCom stamp is irrefutable evidence that CentCom
approved something. Or, alternatively, someone smuggled one aboard!

Tried to resolve a merge conflict.