Fix for #657
/obj/machinery/bot did not implement any behavior on attack_generic, which
is what hostile mobs call on them. The only reason it compiled is because
/obj itself has attack_generic stub with a return in it.

Added an attack_generic proc on bots, and modified hostile code to only
attack 'on' bots. (Why woud a carp care about a thing that looks like a
bucket when it's not moving?)

Also fixed several runtimes with trying to check the digestability of a
bot.

Fixes #657